rstats_requirements.txt
# rstats_requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
rasterio==1.3.10
numpy==2.1.2
shapely==2.0.6
pyproj==3.6.1
